+++ dev_setup.py
@@ -14,7 +14,7 @@ from codecs import encode, decode
 
 ProgramName = "SPIKE"
 VersionName = "Development version"
-VersionInfo = ["0", "99", "1"]   # Major - Minor - Micro
+VersionInfo = ["0", "99", "2  "]   # Major - Minor - Micro
 
 # Major.minor.micro (int) + name eg SPIKE_2_19_5
 # N.M.L
@@ -27,6 +27,15 @@ VersionInfo = ["0", "99", "1"]   # Major - Minor - Micro
 # Release Notes in md syntax !
 release_notes="""
 # Relase Notes
+#### 0.99.2 - January 2019
+- added the .set_unit() method
+- added number of local peaks in bucketing
+- improved Bruker importer and added support for NEO/TopSpin 4.0 files
+- improved importing DOSY processed with TopSpin
+- corrected a bug for min value in bucketing
+- changed pylint/QC defaults -> new values (and corrected a bad bug)
+- cleaned the code a little
+
 #### 0.99.1 - November 2018
 - added the sane algorithm
 - added the pg-sane algorithm