+++ util/mail_error_logger.py
@@ -18,10 +18,18 @@ from __future__ import print_function
 import sys, os
 import threading
 import smtplib
-from email.MIMEMultipart import MIMEMultipart
-from email.MIMEBase import MIMEBase
-from email.MIMEText import MIMEText
-from email import Encoders
+import sys
+if sys.version_info[0] < 3:
+  from email.MIMEMultipart import MIMEMultipart
+  from email.MIMEBase import MIMEBase
+  from email.MIMEText import MIMEText
+  from email import Encoders
+else:
+  from email import encoders as Encoders
+  from email.mime.base import MIMEBase
+  from email.mime.text import MIMEText
+  from email.mime.multipart import MIMEMultipart
+
 from time import sleep, localtime, strftime
 from uuid import getnode as get_mac
 import os.path as op