+++ version_rev.py
@@ -10,8 +10,8 @@ except:
     revision = "--not determined--"
 ProgramName = 'SPIKE'
 VersionName = 'Development version'
-version = '0.99.1'
-rev_date = '26-11-2018'
+version = '0.99.2  '
+rev_date = '17-01-2019'
 
 def report():
     "prints version name when SPIKE starts"